+++ src/main/java/racingcar/view/InputView.java
@@ -11,7 +11,7 @@ public class InputView {
     public List<String> readCarNames() {
         System.out.println("경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)");
         String input = Console.readLine();
-        String[] carNames = input.split(",");
+        String[] carNames = input.split(", ");
         List<String> trimmedCarNames = Arrays.stream(carNames)
                 .map(String::trim)
                 .collect(Collectors.toList());
